@@ -2,7 +2,7 @@
 name = "medea-flutter-webrtc-native"
 version = "0.0.0"
 edition = "2021"
-rust-version = "1.62"
+rust-version = "1.70"
 publish = false
 
 [lib]
