@@ -60,8 +60,9 @@ class _GetUserMediaSampleState extends State<GetUserMediaSample> {
       var stream = await getUserMedia(caps);
       _mediaDevicesList = await enumerateDevices();
       _tracks = stream;
-      await _localRenderer.setSrcObject(
-          _tracks!.firstWhere((track) => track.kind() == MediaKind.video));
+      await _localRenderer.setSrcObject(_tracks!.firstWhere((track) {
+        return track.kind() == MediaKind.video;
+      }));
     } catch (e) {
       print(e.toString());
     }
